server/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "control-deck-backend"
version = "0.0.1"
description = "Control Deck backend (FastAPI + WebSocket)"
authors = [{ name = "Control Deck" }]
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.109.0",
  "uvicorn[standard]>=0.27.0",
  "websockets>=12.0",
  "pydantic>=2.5.0",
  "pydantic-settings>=2.1.0",
  "python-multipart>=0.0.6",
  "aiofiles>=23.2.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "loguru>=0.7.0",
  "pynput>=1.7.6",
  "pyautogui>=0.9.54",
  "pycaw>=20230407",
  "pillow>=10.0.0",
  "psutil>=5.9.0",
  "zeroconf>=0.131.0",
  "qrcode>=7.4.0",
  "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "httpx[http2]>=0.26.0"]

[tool.uvicorn]
factory = false
reload = true
app = "app.main:app"
host = "0.0.0.0"
port = 4455
